pyproject.toml
[project]
name = "mapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aggdraw>=1.3.19",
    "pillow>=12.0.0",
]
